requirements.txt
django
djangorestframework
markdown
pillow
django-filter
requests
django-cors-headers
djangorestframework-jwt

